aws-dashboard-backend-modular/requirements.txt
Flask
flask-cors
boto3
python-dotenv
requests
python-dateutil
Werkzeug
#Flask-SQLAlchemy
#SQLAlchemy